backend/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
defusedxml==0.7.1
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==7.3.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
oauthlib==3.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
PyJWT==2.10.1
python-dotenv==1.1.1
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
